@@ -3,62 +3,53 @@
 	"version": "1.0.0",
 	"private": true,
 	"description": "A frontend React Redux",
-	"main": "index.js",
 	"scripts": {
 		"predeploy": "npm run build",
-		"deploy": "gh-pages -d build",
-		"test": "echo \"Error: no test specified\" && exit 1",
-		"start": "webpack-dev-server --mode development --hot --progress --color --port 3000 --open",
-		"build": "webpack -p --progress --colors",
+		"start-dev": "webpack-dev-server --mode development --hot --progress --color --port 3000 --open",
+		"build": "rimraf dist/ && webpack --mode production --progress",
 		"format": "yarn prettier --write ."
 	},
 	"license": "MIT",
-	"devDependencies": {
-		"@babel/core": "^7.7.7",
-		"@types/classnames": "^2.2.9",
-		"@types/history": "^4.7.3",
-		"@types/node": "^13.11.0",
-		"@types/react": "^16.9.17",
-		"@types/react-dom": "^16.9.4",
-		"@types/react-redux": "^7.1.5",
-		"@types/react-router": "^5.1.5",
-		"@types/redux-actions": "^2.6.1",
-		"@types/webpack": "^4.41.0",
-		"@types/webpack-env": "^1.15.1",
-		"babel-loader": "^8.0.6",
-		"clean-webpack-plugin": "^3.0.0",
-		"css-loader": "^3.4.2",
-		"file-loader": "^6.0.0",
-		"gh-pages": "^2.1.1",
-		"html-loader": "^1.1.0",
-		"html-webpack-plugin": "^4.0.0-beta.11",
-		"mini-css-extract-plugin": "^0.9.0",
-		"postcss": "^7.0.25",
-		"postcss-browser-reporter": "^0.6.0",
-		"postcss-import": "^12.0.1",
-		"postcss-loader": "^3.0.0",
-		"postcss-preset-env": "^6.7.0",
-		"postcss-reporter": "^6.0.1",
-		"postcss-url": "^8.0.0",
-		"prettier": "^2.0.2",
-		"react-hot-loader": "^4.12.18",
-		"style-loader": "^1.1.3",
-		"ts-loader": "^6.2.2",
-		"typescript": "^3.7.4",
-		"url-loader": "^4.0.0",
-		"webpack": "^4.41.5",
-		"webpack-cli": "^3.3.10",
-		"webpack-dev-server": "^3.10.1"
-	},
 	"dependencies": {
+		"@emotion/react": "^11.7.1",
+		"@emotion/styled": "^11.6.0",
+		"@mui/material": "^5.4.0",
 		"@reduxjs/toolkit": "^1.7.1",
-		"classnames": "^2.2.6",
-		"react": "^16.12.0",
-		"react-dom": "^16.12.0",
-		"react-redux": "^7.1.1",
-		"react-router": "^5.1.2",
-		"redux": "^4.0.5",
-		"redux-actions": "^2.6.5",
-		"redux-thunk": "^2.3.0"
+		"classnames": "^2.3.1",
+		"react": "^17.0.2",
+		"react-dom": "^17.0.2",
+		"react-redux": "^7.2.6",
+		"react-router-dom": "^6.2.1",
+		"redux": "^4.1.2",
+		"redux-thunk": "^2.4.1"
+	},
+	"devDependencies": {
+		"@babel/core": "^7.16.12",
+		"@babel/preset-env": "^7.16.11",
+		"@babel/preset-react": "^7.16.7",
+		"@babel/preset-typescript": "^7.16.7",
+		"@pmmmwh/react-refresh-webpack-plugin": "^0.5.4",
+		"@types/node": "^13.11.0",
+		"@types/react": "^17.0.38",
+		"@types/react-dom": "^17.0.11",
+		"@types/react-redux": "^7.1.22",
+		"@types/react-router-dom": "^5.3.3",
+		"@types/webpack": "^5.28.0",
+		"core-js": "^3.21.0",
+		"babel-loader": "^8.2.3",
+		"css-loader": "^6.5.1",
+		"file-loader": "^6.2.0",
+		"html-loader": "^3.1.0",
+		"url-loader": "^4.1.1",
+		"style-loader": "^3.3.1",
+		"html-webpack-plugin": "^5.5.0",
+		"mini-css-extract-plugin": "^2.5.3",
+		"prettier": "^2.5.1",
+		"react-refresh": "^0.11.0",
+		"rimraf": "^3.0.2",
+		"typescript": "4.5.5",
+		"webpack": "^5.68.0",
+		"webpack-cli": "^4.9.2",
+		"webpack-dev-server": "^4.7.3"
 	}
 }
