@@ -1,7 +1,7 @@
 import { createAsyncThunk, ActionReducerMapBuilder } from "@reduxjs/toolkit";
 import { UsersState } from "./state";
 
-const getUserById = createAsyncThunk("users/getUserById", (payload, thunk) => {});
+const getUserById = createAsyncThunk("users/getUserById", async (payload, thunk) => {});
 
 function createCases(builder: ActionReducerMapBuilder<UsersState>) {
 	builder.addCase(getUserById.rejected, (state, action) => {});
