@@ -1,35 +1,29 @@
-var webpack = require("webpack");
-var path = require("path");
-var package = require("./package.json");
+const webpack = require("webpack");
+const path = require("path");
 
-// variables
-var isProduction = process.argv.indexOf("-p") >= 0 || process.env.NODE_ENV === "production";
-var sourcePath = path.join(__dirname, "./src");
-var outPath = path.join(__dirname, "./build");
+const HtmlWebpackPlugin = require("html-webpack-plugin");
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+const ReactRefreshWebpackPlugin = require("@pmmmwh/react-refresh-webpack-plugin");
 
-// plugins
-var HtmlWebpackPlugin = require("html-webpack-plugin");
-var MiniCssExtractPlugin = require("mini-css-extract-plugin");
-var { CleanWebpackPlugin } = require("clean-webpack-plugin");
+const isProduction = process.argv.indexOf("production") >= 0 || process.env.NODE_ENV === "production";
 
+/** @type {import('webpack').Configuration} */
 module.exports = {
-	context: sourcePath,
+	mode: isProduction ? "production" : "development",
+	context: path.resolve(__dirname, "src"),
 	entry: {
-		app: "./main.tsx"
+		app: "./index.tsx"
 	},
 	output: {
-		path: outPath,
-		filename: isProduction ? "[contenthash].js" : "[hash].js",
+		path: path.resolve(__dirname, "dist"),
+		filename: isProduction ? "[name].[contenthash].js" : "[name].[hash].js",
 		chunkFilename: isProduction ? "[name].[contenthash].js" : "[name].[hash].js"
 	},
 	target: "web",
 	resolve: {
 		extensions: [".js", ".ts", ".tsx"],
-		// Fix webpack's default behavior to not load packages with jsnext:main module
-		// (jsnext:main directs not usually distributable es6 format, but es6 sources)
-		mainFields: ["module", "browser", "main"],
 		alias: {
-			app: path.resolve(__dirname, "src/app/")
+			"@app": path.resolve(__dirname, "src", "app")
 		}
 	},
 	module: {
@@ -37,13 +31,20 @@ module.exports = {
 			// .ts, .tsx
 			{
 				test: /\.tsx?$/,
+				exclude: /node_modules/,
 				use: [
-					!isProduction && {
+					{
 						loader: "babel-loader",
-						options: { plugins: ["react-hot-loader/babel"] }
-					},
-					"ts-loader"
-				].filter(Boolean)
+						options: {
+							plugins: [!isProduction && "react-refresh/babel"].filter(Boolean),
+							presets: [
+								["@babel/preset-env", { useBuiltIns: "usage", corejs: "3.21" }],
+								"@babel/preset-typescript",
+								"@babel/preset-react"
+							]
+						}
+					}
+				]
 			},
 			// css
 			{
@@ -51,32 +52,7 @@ module.exports = {
 				use: [
 					isProduction ? MiniCssExtractPlugin.loader : "style-loader",
 					{
-						loader: "css-loader",
-						query: {
-							sourceMap: !isProduction,
-							importLoaders: 1,
-							modules: {
-								localIdentName: isProduction ? "[hash:base64:5]" : "[local]__[hash:base64:5]"
-							}
-						}
-					},
-					{
-						loader: "postcss-loader",
-						options: {
-							ident: "postcss",
-							plugins: [
-								require("postcss-import")({ addDependencyTo: webpack }),
-								require("postcss-url")(),
-								require("postcss-preset-env")({
-									/* use stage 2 features (defaults) */
-									stage: 2
-								}),
-								require("postcss-reporter")(),
-								require("postcss-browser-reporter")({
-									disabled: isProduction
-								})
-							]
-						}
+						loader: "css-loader"
 					}
 				]
 			},
@@ -91,7 +67,6 @@ module.exports = {
 	},
 	optimization: {
 		splitChunks: {
-			name: true,
 			cacheGroups: {
 				commons: {
 					chunks: "initial",
@@ -112,10 +87,8 @@ module.exports = {
 			NODE_ENV: "development", // use 'development' unless process.env.NODE_ENV is defined
 			DEBUG: false
 		}),
-		new CleanWebpackPlugin(),
 		new MiniCssExtractPlugin({
-			filename: "[hash].css",
-			disable: !isProduction
+			filename: "[hash].css"
 		}),
 		new HtmlWebpackPlugin({
 			template: "assets/index.html",
@@ -126,33 +99,13 @@ module.exports = {
 				useShortDoctype: true,
 				collapseWhitespace: true,
 				collapseInlineTagWhitespace: true
-			},
-			append: {
-				head: `<script src="//cdn.polyfill.io/v3/polyfill.min.js"></script>`
-			},
-			meta: {
-				title: package.name,
-				description: package.description,
-				keywords: Array.isArray(package.keywords) ? package.keywords.join(",") : undefined
 			}
-		})
-	],
+		}),
+		!isProduction && new ReactRefreshWebpackPlugin()
+	].filter(Boolean),
 	devServer: {
-		contentBase: sourcePath,
-		hot: true,
-		inline: true,
-		historyApiFallback: {
-			disableDotRule: true
-		},
-		stats: "minimal",
-		clientLogLevel: "warning"
+		static: path.resolve(__dirname, "src"),
+		hot: true
 	},
-	// https://webpack.js.org/configuration/devtool/
-	devtool: isProduction ? "hidden-source-map" : "cheap-module-eval-source-map",
-	node: {
-		// workaround for webpack-dev-server issue
-		// https://github.com/webpack/webpack-dev-server/issues/60#issuecomment-103411179
-		fs: "empty",
-		net: "empty"
-	}
+	devtool: isProduction ? "hidden-source-map" : "cheap-module-source-map"
 };
